feat: 查询订单前排队接口、订单保存预期等待时间 todo: 下单后前端显示前面排队状态

+++ hutu-model/src/main/java/org/nott/model/BizItem.java
@@ -81,9 +81,9 @@ public class BizItem implements Serializable {
     private Date updateTime;
 
     /**
-     * 预计制作时长，单位：秒
+     * 预计制作时长，单位：分
      */
-    private Long expectMakeTime;
+    private Integer expectMakeTime;
 
 
 }
